Added a separate timeout parameter for vrrp_script checks
Merge pull request from Jonathan Harden:

I've added a timeout parameter to the vrrp check scripts which
allow you to have the check timeout different to the interval.
When no timeout has been specified the interval is used (which
mimics the current behaviour).

To explain the reasoning: We wanted to have check scripts time
out faster than our check interval. Doing the check we need to
perform is a little load intensive and so we don't want to
perform it every few seconds. With this patch we set an
interval of 60 seconds but a timeout of 5 seconds (if the check
takes more than a few seconds then the service is not working
correctly).