Add support to LVS One-Packet Scheduling (known as OPS)
This patch comes from Erik de Groot, <edg1972 <at> gmail <dot> com>. Thanks for your work Erik !

Discussion from Erik:

Typically RADIUS traffic comes from a limited amount of clients and thus you have a very limited
range of IP tuples in action which will never expire. Issue with Keepalived without this patch is
that, although it correctly re-assigns traffic when a real server dies, it will never re-assign
traffic back to the real server when it is restored. This is because LVS creates virtual
connections, for each IP tuple, that will never time out as the clients keep sending traffic to
the server.

With this patch is is possible to enable OPS for UDP virtual servers which means LVS does not
create virtual connections and takes a new loadbalancing decision for each UDP packet. The result
is that a restored server now gets RADIUS traffic as soon as LVS has taken it it back into the
server pool.