Fix VRRP wrong interface flags corner case
If a link event arrives between the initial scanning for interfaces and
configuration file parsing, the VRRP instance will enter an unrecoverable
state. This fix will update the interface flags even when the interface
exists, not only for the inital scan.
Note that when all is up and running the link events will be properly
handled by netlink, so this fix only fixes the special case when a link
changes state during initalization/configuration.

Signed-off-by: Jonas Johansson <jonasj76@gmail.com>