vrrp: Support xmit VRRP packets from base VMAC interface.
Here is a merge of patch from Olipro (Oliver Smith). I just reworked a
little initial shot. Thanks for your job and idea in here Oliver.

Here are comments from Olipro :

This provides a new option to use in conjunction with the VMAC
functionality which will result in VRRP advertisements being sent and
received over the underlying interface (and therefore having the source
MAC of that interface rather than the VMAC device)

With this new functionality enabled, VRRP messages will not affect the
switch MAC address table since the non-unique VMAC address is now used
only for sending a gratuitous ARP, thereby ensuring that in conditions
of VRRP message loss, a probing partner will not inadvertently take over
traffic.

This also resolves issues where VRRP messages are not successfully being
seen on the VMAC interface as with the new option, the underlying
interface is also used to listen out for VRRP messages.