unify connection options among checkers
All the remote checkers (TCP, HTTP/SSL, SMTP) now have the same set of connection options:
 * connect_ip (new to TCP, HTTP)
 * connect_port
 * bindto
 * bind_port  (new)
 * connect_timeout (new to SMTP)

All of them are optional with reasonable defaults.
The patch is designed for simplicity in adding a new option.

Since the connect_ip could be inequal to the RS address and, worse, the same for all RSes,
the endpoint is now logged as [RS]:rport, not the [connect_ip]:connect_port.