Fix mismatched advertisement interval
In VRRP version 3, all BACKUP routers must set their advertisement
intervals to match the current MASTER's. Although not explicitly
stated in RCF5798, when the MASTER falls over or forfeits its
MASTER status, the new MASTER should not continue to use the old
MASTER's advertisement interval value and should instead use its
locally configured value.

To achieve this, a new field has been added to the VRRP structure
that stores the most recent advertisement interval of the current
MASTER. We track changes to the current MASTER's interval and update
this new variable accordingly. The value is only updated when we are
in BACKUP state and reconfiguring the local advertisement interval
has no effect on it.

Signed-off-by: Colin Docherty <cdochert@brocade.com>