ipvs: respect the error code of the ipvs_talk
Previously, if the IPVS reflector was unable to perform
its task, it reported error through syslog and ignored it.

This behavior leads to inconsistancies with quorum-handler:
it is called with UP even if no RS were added into the IPVS.
This could take place, for example, when there is a limit of
opened filehandles and keepalived was unable to open netlink
socket (it is opened on every call to the ipvs_talk).

Now the check is not marked as OK unless IPVS reflector reports
OK. Following successfull check will try to add an RS again.

The special case errors "ENOENT on remove" and "EEXIST on add"
are treated with OK result code.