ipvs: new service option "ip_family"
This option explicitly specifies the address family of a
fwmark IPVS service entry. Previously it was determined by
the AF of the first real server. This logic is kept as a fallback
when the "ip_family" option is missing.

Also, now it is possible to create two different services
for v4 and v6 with the same fwmark number.