Set (and restore) interface parameters
In order to receive and send multicasts on the correct interfaces
various parameters need to be set via the /proc/sys/net/ipv4/conf
interface. This patch sets them as needed, and restores any
changes on the underlying interface on exit.

If a user currently sets any parameters by scripts, that will
override these changes and still work, but this change in general
will make it unnecessary to change any parameters with scripts.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>