Allow gratuitious ARP parameters to be configured globally
It is likely that the gratuitions ARP parameters will want to be
the same for all interfaces, so allow the defaults to be set
globally.

Also allow vrrp_garp_delay to be set to 0 to indicate not to
send further garp messages after a delay (to emulate how the
kernel sends gratuitous ARPs).

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>