Make parent process handle and propagate USR1/2 signals
In order to be able to automate writing configuration and/or stats
the signals USR1 and USR2 need to be able to be sent to the parent
process since its pid can be read from /var/run/keepalived.pid.
The parent then needs to propagate these signals to a vrrp child.

Resolves issue #169.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>