Return address of previous signal handler according to SA_SIGINFO
The man page for sigaction(2) states that SA_SIGINFO is only
meaningful when establishing a signal handler. This appears not
to be the case, since the flag will be set in the oldact structure
on return from sigaction if the previous signal handler was
established using the SA_SIGINFO flag.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>