Don't wait on script process being killed after timeout
The child_timout_thread functions send a SIGKILL to a child
process that has timed out and didn't die quickly enough
after sending a SIGTERM. They then wait on the process dying.

The main problem is that if the waitpid is successful here, then
waitpid in thread_child_handler will never be successful for the
same pid, and so the entry on the child list will never be removed
and the parent thread will not be marked as ready.

There is also a theoretical possibility that the child process is
unkillable, and so the waitpid would hang forever.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>