Move common code for opening fd 0/1/2 into a function
The code for setting fd 0/1/2 to /dev/null before running a script
was in several places. All the common code is moved into a function
and the function called from the relevant places.

It is only necessary to reopen fd 0/1/2 if keepalived is running
with the --dont-fork option, since without that option the fds are
already open on /dev/null.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>